requirements.txt
Django==5.2.11
python-dotenv==1.2.1
Pillow==11.1.0
dj-database-url==2.2.0
psycopg2-binary==2.9.10
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
django-allauth==65.3.0
PyJWT==2.9.0
cryptography==44.0.0
